dts: hikey: set pinmux based on module
In the previous dts, it will set pinmux for every pin one by one; it's
not easily used by the modules and difficult to maintain them.

So change to set pinmux based on module, so that we can group pins base
on that. And every module can easily hook the pinmux setting.

Signed-off-by: Leo Yan <leo.yan@linaro.org>