drm/hisilicon: Remove the old 720p canned mode
Since the old 720p is not work as it should be. That's to say we can not
reproduce the old 720p as before. Becuase many changes has been change(hdmi,
clock) since Andy's patches.

This also fix the bug 92 by not using adjust mode:
Adjust mode will be free, so we cannot use it in dpms operation again after a
mode has been set.

Signed-off-by: Xinliang Liu <xinliang.liu@linaro.org>