[hack]driver: misc: hi6220-sysconfig: enable CoreSight
By configuring a global register to enable CoreSight, if we don't set
this register, the CoreSight ETM cannot be initialized, but other
components of CoreSight can be even without this configuration.

Moreover, this configuration should put into UEFI, because the register
isn't suppose to be set other values while kernel running.

Signed-off-by: Fei Wang <w.f@huawei.com>
Signed-off-by: Chunyan Zhang <zhang.chunyan@linaro.org>