dma hisi fix nobody cared as soon as any error
As it was before, as soon as the DMAC IP felt there was an error
he would return IRQ_NONE since no actual transfer had completed.

After spinning on that for 100K interrupts, Linux yanks the IRQ with
a "nobody cared" error.

This patch lets it handle the interrupt and keep the IRQ alive.

Signed-off-by: Andy Green <andy.green@linaro.org>
[jstultz: Forward ported to 4.1]
Signed-off-by: John Stultz <john.stultz@linaro.org>