adv7511: Fix use of drm_display_mode pointer after free
The adv7511 structure uses a drm_display_mode pointer,
which is assigned to a value allocated higher in the
call stack that is later destryed.

This results in the adv7511 initializing occasionally
with garbage values.

Fix this issue by using a drm_disply_mode structure
instead of a pointer in the adv7511 structure, and
copying the provided structure over entirely.

Signed-off-by: John Stultz <john.stultz@linaro.org>