Revert "Quick fix so that driver retrieves addresses from device tree instead of using hardcoded defines."
This reverts commit c9549165f4bdf66cba3c5edecfee91a7bfa829fa.

I just saw yesterday that I made a small mistake with the if sentances. The function of_address_to_resource returns 0 (zero) on sucess and -EINVAL upon failure meaning the error message is printed even though everything is setup correctly. The exclamation mark should be removed in the if sentances at line 99 and line 107, i.e. change to:

99: if(of_address_to_resource(peri_node, 0, &res)) {

107: if(of_address_to_resource(pm_node, 0, &res)) {

Sorry about that!
Regards,
Erik