drm/i2c: adv7511: Use internal timing generator
ADV7533 provides an internal timing generator for certain modes that it
can't use the DSI clock directly.

We've observed that hdmi is more stable with the internal timing
generator, especially if there are instabilities in the dsi clock source.

The data spec recommends usage of the timing generator too. Create a DT
parameter that allows us to enable/disable the timing generator.

Signed-off-by: Archit Taneja <architt@codeaurora.org>